authnication/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
numpy==2.2.1
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
